+++ src/Extensions/ArrayExtension.php
@@ -46,8 +46,8 @@ class ArrayExtension extends Service
     /**
      * @inheritDoc
      */
-    public function __invoke(ContainerInterface $c, $prev = [])
+    public function __invoke(ContainerInterface $c, array $prev = [])
     {
-        return array_merge($prev, $this->resolveKeys($c, $this->dependencies));
+        return array_merge($prev, $this->resolveDeps($c, $this->dependencies));
     }
 }