+++ src/Factories/FuncService.php
@@ -56,10 +56,11 @@ class FuncService extends Service
 
     /**
      * @inheritDoc
+     * @return callable
      */
     public function __invoke(ContainerInterface $c)
     {
-        $deps = $this->resolveKeys($c, $this->dependencies);
+        $deps = $this->resolveDeps($c, $this->dependencies);
 
         /**
          * @psalm-suppress MissingClosureReturnType Cannot declare mixed until PHP 8