+++ src/Factories/StringService.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Dhii\Services\Factories;
 
+use Dhii\Services\ResolveKeysCapableTrait;
 use Dhii\Services\Service;
 use Psr\Container\ContainerInterface;
+use UnexpectedValueException;
 
 /**
  * A factory for string values. Supports interpolation with dependent service values.
@@ -21,9 +23,13 @@ use Psr\Container\ContainerInterface;
  *      ]),
  *  ]
  *  ```
+ *
+ * @psalm-import-type ServiceRef from Service
  */
 class StringService extends Service
 {
+    use ResolveKeysCapableTrait;
+
     /** @var string */
     protected $format;
 
@@ -43,6 +49,30 @@ class StringService extends Service
     }
 
     /**
+     * Retrieve a service _as a string_ with the specified name from the given container.
+     *
+     * @param ServiceRef $serviceRef The service, or its name in the container..
+     * @param ContainerInterface $c The container to retrieve the service from.
+     *
+     * @return string The string representation of the service.
+     *
+     * @throws UnexpectedValueException If service could be converted to string.
+     */
+    protected function resolveString($serviceRef, ContainerInterface $c): string
+    {
+        $service = $this->resolveSingleDep($c, $serviceRef);
+
+        if (!is_null($service) && !is_scalar($service) && !is_object($service)) {
+            throw new UnexpectedValueException(sprintf(
+                'Service must be of type null|scalar|object to be stringable; %1$s received',
+                gettype($service)
+            ));
+        }
+
+        return strval($service);
+    }
+
+    /**
      * @inheritDoc
      */
     public function __invoke(ContainerInterface $c)
@@ -54,7 +84,7 @@ class StringService extends Service
         $replace = [];
         foreach ($this->dependencies as $idx => $dependency) {
             $idx = (string) $idx;
-            $replace['{' . $idx . '}'] = strval($c->get($dependency));
+            $replace['{' . $idx . '}'] = $this->resolveString($dependency, $c);
         }
 
         return strtr($this->format, $replace);